backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx==0.24.1
supabase>=1.0.3
asyncpg>=0.27.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
databases[postgresql]>=0.9.0
email-validator>=2.0.0
# Pin these to avoid Python 3.12 compatibility issues
hyperframe==6.0.1
h2==4.1.0 